implementing proto fajl za order

+++ src/main/java/rs/raf/pds/v5/z2/gRPC/BidOfferOrBuilder.java
@@ -18,9 +18,9 @@ public interface BidOfferOrBuilder extends
       getSymbolBytes();
 
   /**
-   * <code>int32 price = 2;</code>
+   * <code>double price = 2;</code>
    */
-  int getPrice();
+  double getPrice();
 
   /**
    * <code>int32 numberOfOffers = 3;</code>
